SDL3: Fix window-drag tracking under 2x display scaling
Regression from the SDL2->SDL3 port: SDL3 mouse coordinates are float,
and under 2x logical scaling the pump's divide-by-scale yields .5-valued
coords. WindowFrameWidget stored the drag baseline in an int and did
'm_last_mouse_x = event.x', truncating the fractional remainder each
event; combined with the truncated int delta this systematically dropped
half-pixel steps and stalled left/up drags (SDL2 coords were already
integers, so the bug was latent). Keep the drag/resize baselines in
float and advance the baseline only by the integer amount emitted so the
sub-pixel remainder carries to the next motion.

+++ include/widget/windowing/WindowFrameWidget.h
@@ -305,8 +305,11 @@ private:
     
     // Drag state
     bool m_is_dragging;
-    int m_last_mouse_x;
-    int m_last_mouse_y;
+    // Float: SDL3 mouse coordinates are float and become non-integer under
+    // display scaling (e.g. 2x). Keeping the drag baseline in float and
+    // carrying the fractional remainder avoids truncation stalling the drag.
+    float m_last_mouse_x;
+    float m_last_mouse_y;
     
     // Double-click detection for close
     Uint32 m_last_click_time;
@@ -315,8 +318,8 @@ private:
     // Resize state
     bool m_is_resizing;
     int m_resize_edge; // 0=none, 1=left, 2=right, 4=top, 8=bottom, combinations for corners
-    int m_resize_start_x;
-    int m_resize_start_y;
+    float m_resize_start_x;  // float: see m_last_mouse_x
+    float m_resize_start_y;
     int m_resize_start_width;
     int m_resize_start_height;
     int m_resize_start_window_x;
